feat: Add option to disable URL unfurling in Slacker
Add a feature to allow users to choose not to unfurl a URL or link when using Slacker. This is useful for cases where unfurling is not desired or necessary. Additionally, a test case for this feature has been added to main.go.

Closes #11

unfurling option not working properly

fix: typo for greptile

feat: slacker now handles unfurling option + main.go for testing purpose

fix: main.go now uses a youtube URL